seed mongodb as script with pnpm
