+++ package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "start": "next start",
     "storybook": "storybook dev -p 6006",
-    "build-storybook": "storybook build"
+    "build-storybook": "storybook build",
+    "seed": "node ./app/api/mongodb/seed_mongodb.js"
   },
   "dependencies": {
     "@heroicons/react": "^2.1.4",
